optimization/requirements.txt
fal-client>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
polars>=0.20.0
pandas>=1.5.0
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0
pillow>=9.0.0
tqdm>=4.64.0
numpy>=1.21.0
aesthetic-predictor-v2-5
hpsv2
dspy-ai